Split e2e package to each e2e test
